Integrate pbrt parsing and pandora scene file output into a single script

+++ projects/pbrt_importer/python/export_pandora.py
@@ -1,9 +1,10 @@
 import argparse
-import pickle
-import ujson # ujson is way faster than the default json module for exporting large dictionaries
+import os
+import parsing
 from config_parser import ConfigParser
 from scene_parser import SceneParser
-import os
+import ujson  # ujson is way faster than the default json module for exporting large dictionaries
+
 
 def extract_pandora_data(pbrt_data, out_mesh_folder):
     if not os.path.exists(out_mesh_folder):
@@ -21,17 +22,28 @@ if __name__ == "__main__":
     parser = argparse.ArgumentParser(
         description="Export preprocessed PBRT scenes to Pandora's JSON format")
     parser.add_argument("--file", type=str,
-                        help="Path to preprocessed PBRT scene (binary) file")
+                        help="Path to PBRT scene file")
     parser.add_argument("--out", type=str,
-                        help="Name of output scene description file")
+                        help="Name/path of output Pandora scene description file")
     args = parser.parse_args()
+
+    if not os.path.isfile(args.file):
+        print("Error: pbrt scene file not found")
+        exit(1)
+    if not os.path.exists(os.path.dirname(args.out)):
+        print("Error: output path does not exist")
+        exit(1)
 
     out_mesh_folder = os.path.join(os.path.dirname(args.out), "pandora_meshes")
 
-    pbrt_data = pickle.load(open(args.file, "rb"))
+    #pbrt_data = pickle.load(open(args.file, "rb"))
+    print("==== PARSING PBRT FILE ====")
+    pbrt_data = parsing.parse_file(args.file)
+
+    print("==== CONVERTING TO PANDORA FORMAT ====")
     pandora_data = extract_pandora_data(pbrt_data, out_mesh_folder)
 
-    print("Writing JSON")
+    print("==== WRITING PANDORA SCENE FILE ====")
     ujson.dump(pandora_data, open(args.out, "w"), indent=2)
